chore(artist): renames artist header
