+++ src/Apps/Artist/Routes/Overview/Components/ArtistCareerHighlights.tsx
@@ -11,7 +11,7 @@ import { ArtistCareerHighlightsQuery } from "__generated__/ArtistCareerHighlight
 import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
 import { FC } from "react"
 import { RailHeader } from "Components/Rail/RailHeader"
-import { ARTIST_HEADER_NUMBER_OF_INSIGHTS } from "Apps/Artist/Components/ArtistHeader/ArtistHeader2"
+import { ARTIST_HEADER_NUMBER_OF_INSIGHTS } from "Apps/Artist/Components/ArtistHeader/ArtistHeader"
 import { ArtistCareerHighlightFragmentContainer } from "Apps/Artist/Routes/Overview/Components/ArtistCareerHighlight"
 import { Media } from "Utils/Responsive"
